Finished refactoring component names
