+++ src/views/TheAboutPage.vue
@@ -44,7 +44,7 @@
 
 <script>
 export default {
-  name: 'About',
+  name: 'TheAboutPage',
   data () {
     return {
       contributors: ['Frank Matranga', 'Christie Nero', 'Abigail Medina', 'Bowen Wang', 'Ziyi Wang', 'Seth Laurenceau', 'Zach Love', 'Kikiola Sanusi', 'Tyler Machado']
