+++ src/views/TheDashboard.vue
@@ -19,7 +19,7 @@ import { FullCalendar } from 'vue-full-calendar';
 import 'fullcalendar/dist/fullcalendar.css';
 
 export default {
-  name: 'Dashboard',
+  name: 'TheDashboard',
   components: { FullCalendar },
   data () {
     return {
